move snplocus query code out of regression.js
